# docs/swagger.yaml
definitions:
  api_key.CreateAPIKeyRequest:
    properties:
      app_description:
        maxLength: 100
        type: string
      app_homepage_url:
        maxLength: 100
        type: string
      app_name:
        maxLength: 20
        type: string
      notify_url:
        maxLength: 100
        type: string
      redirect_uri:
        maxLength: 100
        type: string
    required:
    - app_homepage_url
    - app_name
    - notify_url
    type: object
  api_key.UpdateAPIKeyRequest:
    properties:
      app_description:
        maxLength: 100
        type: string
      app_homepage_url:
        maxLength: 100
        type: string
      app_name:
        maxLength: 20
        type: string
      notify_url:
        maxLength: 100
        type: string
      redirect_uri:
        maxLength: 100
        type: string
    type: object
...
  util.ResponseAny:
    properties:
      data: {}
      error_msg:
        example: ""
        type: string
    type: object
info:
  contact: {}
  title: LINUX DO PAY
  version: 0.1.0
paths:
...
  /api/v1/merchant/api-keys:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - merchant
    post:
      consumes:
      - application/json
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api_key.CreateAPIKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - merchant
  /api/v1/merchant/api-keys/{id}:
    delete:
      parameters:
      - description: API Key ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - merchant
    get:
      parameters:
      - description: API Key ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - merchant
    put:
      consumes:
      - application/json
      parameters:
      - description: API Key ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api_key.UpdateAPIKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - merchant
  /api/v1/merchant/api-keys/{id}/payment-links:
    get:
      parameters:
      - description: API Key ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - merchant
    post:
      consumes:
      - application/json
      parameters:
      - description: API Key ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - merchant
  /api/v1/merchant/api-keys/{id}/payment-links/{linkId}:
    delete:
      parameters:
      - description: API Key ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: Payment Link ID
        format: int64
        in: path
        name: linkId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - merchant
...
swagger: "2.0"